Allow IOCTLs and mmaps backed by host devices
- The ioctl() syscalls on device-backed allowed file descriptors are
  pass-through. On the Linux-SGX PAL, the allowed IOCTL requests must
  be explicitly allowed in the manifest via the new option
  `sgx.allowed_ioctls.[id].request`. Also on the Linux-SGX PAL, the
  allowed IOCTLs' arguments (typically pointers to complex nested
  objects) must be explicitly described in the manifest via the new
  options `sgx.ioctl_structs.[id]` and a corresponding reference
  `sgx.allowed_ioctls.[id].struct`; see docs for explanation of
  the IOCTL struct format.
- The mmap() syscalls on device-backed allowed file descriptors are
  pass-through. On the LinuxSGX PAL, all mmapped memory regions reside
  in untrusted memory, and the enclave logic must be careful when
  accessing it (typically the app developer must use this mmapped
  memory region as an encrypted bounce buffer).

Device-backed ioctl() and mmap() syscalls are insecure by themselves
since they only pass the arguments to and from the untrusted memory.
It is the responsibility of the app developer to correctly use them
for secure communication with the host devices.

This commit also adds a hack to Graphene-SGX: if the application
requests `mmap(MAP_ANONYMOUS | MAP_SHARED)` under Linux-SGX PAL,
Graphene considers this a request to allocate untrusted memory (e.g.,
for a buffer that will be later shared with a host device).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>